A data frame contains the details of the transformation of the metabolic reconstruction

A data frame contains the details of the transformation of the metabolic reconstruction

Usage

conversionMatrix

Format

A data frame contains 93 rows and 67 variables:

MarkerWords

the key words of the steps of metabolic reconstruction

ThresholdValue

the values filter the steps after the mapping

Steps

the steps of metabolic reonstruction

Group

the classification of the steps

SpeciesName

1 means the output of Species name.

TaxonID

1 means the output of Taxon ID.

NCBI

1 means the output of NCBI.

Uniprot

1 means the output of Uniprot.

KEGG

1 means the output of KEGG.

MetaCyc

1 means the output of MetaCyc.

PubChem

1 means the output of PubChem.

Brenda

1 means the output of Brenda.

Download

1 means the output of Download.

GeneticInformation

1 means the output of Genetic information.

ProteinInformation

1 means the output of Protein information.

GenomeSequence

1 means the output of Genome sequence.

ProteinSequence

1 means the output of Protein sequence.

MetabolicFunctionInformation

1 means the output of Metabolic function information.

Metabolites

1 means the output of Metabolites.

Cofactor

1 means the output of Cofactor.

Nucleotides

1 means the output of Nucleotides.

AminoAcid

1 means the output of Amino acid.

MolecularWeight

1 means the output of Molecular weight.

DryWeight

1 means the output of Dry weight.

MetabolicReaction

1 means the output of Metabolic reaction.

TerminalMetabolite

1 means the output of Terminal metabolite.

Secretion

1 means the output of Secretion.

BiomassReaction

1 means the output of Biomass reaction.

DemandReaction

1 means the output of Demand reaction.

SinkReaction

1 means the output of Sink reaction.

GapReaction

1 means the output of Gap reaction.

SpontaneousReaction

1 means the output of Spontaneous reaction.

ExtracellularAndPeriplasmicTransportReactions

1 means the output of Extracellular and periplasmic transport reactions reaction.

ExchangeReaction

1 means the output of Exchange reaction.

IntracellularTransportReaction

1 means the output of Intracellular transport reaction.

ReactionFlux

1 means the output of Reaction flux.

GPR

1 means the output of GPR.

BlastComparison

1 means the output of Blast comparison.

Homology

1 means the output of Homology.

HomologousGene

1 means the output of Homology gene.

StoichiometricMatrix

1 means the output of Stoichiometric matrix.

Knockout

1 means the output of Knockout.

TargetReaction

1 means the output of Target reaction.

Restrictions

1 means the output of Restrictions.

GrowthConditions

1 means the output of Growth conditions.

MinORMax

1 means the output of min | max.

ReconstructionData

1 means the output of Reconstruction data.

FVA

1 means the output of FVA.

MetabolicFlux

1 means the output of Metabolic flux.

Statistics

1 means the output of Metabolic Statistics.

NeutralMolecularFormula

1 means the output of Neutral molecular formula.

ChargedMolecularFormula

1 means the output of Charged molecular formula.

LiteratureDataCollection

1 means the output of Literature data collection.

ConservationOfMassAndCharge

1 means the output of Conservation of mass and charge.

GibbsFreeEnergy

1 means the output of Gibbs free energy.

CellCompartmentInformation

1 means the output of Cell compartment information.

SubsystemInformation

1 means the output of Subsystem information.

MetaboliteIdentification

1 means the output of Metabolite Identification.

Unite

1 means the output of Unite.

ManualPlanning

1 means the output of Manual planning.

Coefficient

1 means the output of Coefficient.

ScatterPlot

1 means the output of Scatter plot.

TestReport

1 means the output of Test Report.

TypeIIIPath

1 means the output of Type III path.

SBML

1 means the output of SBML file.

Mat

1 means the output of Mat file.

Excel

1 means the output of Excel file.