A data frame produced by the function map_word_to_step

A data frame produced by the function map_word_to_step

Usage

matrixProcess

Format

A data frame contains 67 variables

degree

the number of the steps used in a article.

SpeciesName

-1 means the input of Species name. 1 means the output of Species name.

TaxonID

-1 means the input of Taxon ID. 1 means the output of Taxon ID.

KEGG

-1 means the input of KEGG reaction. 1 means the output of KEGG reaction.

NCBI

-1 means the input of NCBI gene. 1 means the output of NCBI gene.

Uniprot

-1 means the input of Uniprot gene. 1 means the output of Uniprot.

MetaCyc

-1 means the input of MetaCyc gene. 1 means the output of MetaCyc.

PubChem

-1 means the input of PubChem molecular formula. 1 means the output of PubChem molecular formula.

Brenda

identify whether input the Brenda. -1 means the input of Brenda. 1 means the output of Brenda.

Chebi

-1 means the input of Chebi information. 1 means the the output of Chebi information.

WebGCM

-1 means the input of Web GCM. 1 means the the output of Web GCM.

SpontaneousReaction

-1 means the input of spontaneous reaction. 1 means the the output of spontaneous reaction.

ExtracellularAndPeriplasmicTransportReactions

-1 means the input of Extracellular and periplasmic transport reactions. 1 means the the output of Extracellular and periplasmic transport reactions.

ExchangeReaction

-1 means the input of exchange reaction. 1 means the the output of exchange reaction.

MissingExchangeReaction

-1 means the input of missing exchange reaction. 1 means the the output of missing exchange reaction.

IntracellularTransportReaction

-1 means the input of intracellular transport reaction. 1 means the the output of intracellular transport reaction.

Gene

-1 means the input of gene. 1 means the the output of gene.

Protein

-1 means the input of protein. 1 means the the output of protein.

Knockout

-1 means the input of knockout. 1 means the the output of knockout.

StoichiometricMatrix

-1 means the input of stoichiometric matrix. 1 means the the output of stoichiometric matrix.

ObjectiveReaction

-1 means the input of objective reaction. 1 means the the output of objective reaction.

Constraints

-1 means the input of constraints. 1 means the the output of constraints.

Secretion

-1 means the input of secretion. 1 means the the output of secretion.

Mutisecretion

-1 means the input of mutisecretion. 1 means the the output of mutisecretion.

RichMedia

-1 means the input of rich media. 1 means the the output of rich media.

GenomeSequence

-1 means the input of genome sequence. 1 means the the output of genome sequence.

ProteomeSequence

-1 means the input of proteome sequence. 1 means the the output of proteome sequence.

AminoAcidWeight

-1 means the input of amino acid weight. 1 means the the output of amino acid weight.

AminoAcidMolecularWeight

-1 means the input of amino acid molecular weight. 1 means the the output of amino acid molecular weight.

NucleotideWeight

-1 means the input of nucleotide weight. 1 means the the output of nucleotide weight.

NucleotideMolecularWeight

-1 means the input of nucleotide molecular weight. 1 means the the output of nucleotide molecular weight.

DryWeight

-1 means the input of dry weight. 1 means the the output of dry weight.

BiomassReaction

-1 means the input of biomass reaction. 1 means the the output of biomass reaction.

DemandReaction

-1 means the input of demand reaction. 1 means the the output of demand reaction.

SinkReaction

-1 means the input of sink reaction. 1 means the the output of sink reaction.

GapReaction

-1 means the input of gap reaction. 1 means the the output of gap reaction.

MinORMax

-1 means the input of gap reaction. 1 means the the output of gap reaction.

GeneticInformation

-1 means the input of genetic information. 1 means the the output of genetic information.

MetabolicFunction

-1 means the input of metabolic function. 1 means the the output of metabolic function.

Metabolites

-1 means the input of metabolites. 1 means the the output of metabolites.

BiomassMetabolites

-1 means the input of biomass metabolites. 1 means the the output of biomass metabolites.

MetabolicReaction

-1 means the input of metabolic reaction. 1 means the the output of metabolic reaction.

ReconstructionData

-1 means the input of reconstruction data. 1 means the the output of reconstruction data.

DataStatistics

-1 means the input of data statistics. 1 means the the output of data statistics.

NeutralMolecularFormula

-1 means the input of neutral molecular formula. 1 means the the output of neutral molecular formula.

ChargedMolecularFormula

-1 means the input of charged molecular formula. 1 means the the output of charged molecular formula.

ConservationOfMassAndCharge

-1 means the input of conservation of mass and charge. 1 means the the output of conservation of mass and charge.

GibbsFreeEnergyInformation

-1 means the input of gibbs free energy information. 1 means the the output of gibbs free energy information.

Mass_ChargeConservationAssessment

-1 means the input of mass-charge conservation assessment. 1 means the the output of conservation assessment.

CellularCompartment

-1 means the input of cellular compartment. 1 means the the output of cellular compartment.

SubsystemInformation

-1 means the input of subsystem information. 1 means the the output of subsystem information.

IdentifiersInKEGG

-1 means the input of identifiers in KEGG. 1 means the the output of identifiers in KEGG.

IdentifiersInMetaCyc

-1 means the input of identifiers in MetaCyc. 1 means the the output of identifiers in MetaCyc.

UniformIdentifier

-1 means the input of uniform identifier. 1 means the the output of uniform identifier.

Coefficient

-1 means the input of coefficient. 1 means the the output of coefficient.

ScatterPlot

-1 means the input of scatter plot. 1 means the the output of scatter plot.

TerminalMetabolite

-1 means the input of terminal metabolite. 1 means the the output of terminal metabolite.

TypeIIIPathway

-1 means the input of Type III pathway. 1 means the the output of Type III pathway.

NetworkGap

-1 means the input of network gap. 1 means the the output of network gap.

Growth

-1 means the input of growth. 1 means the the output of growth.

BlockReaction

-1 means the input of block reaction. 1 means the the output of block reaction.

MetabolicFlux

-1 means the input of metabolic flux. 1 means the the output of metabolic flux.

ModelPredictCorrectly

1 means the output of model PredictCorrectly.

ModelGrowingTooFast

-1 means the input of model growing too fast. 1 means the the output of model growing too fast.

SBML

-1 means the input of SBML file. 1 means the the output of SBML file.

Mat

-1 means the input of Mat file. 1 means the the output of Mat file.

Excel

-1 means the input of Excel file. 1 means the the output of Excel file.